requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
certifi==2021.10.8
cups==0.0.6
escpos==1.9
future==0.18.2
Pillow==9.1.0
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Quartz==8.5
pyPdf==1.13
pyserial==3.5
pystray==0.19.3
python-barcode==0.13.1
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
viivakoodi==0.8.0
altgraph==0.17.2
appdirs==1.4.4
argcomplete==2.0.0
certifi==2021.10.8
configparser==5.2.0
cups==0.0.6
escpos==1.9
future==0.18.2
macholib==1.16
Pillow==9.1.0
pyinstaller==5.0
pyinstaller-hooks-contrib==2022.4
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Quartz==8.5
pyPdf==1.13
pyserial==3.5
pystray==0.19.3
python-barcode==0.13.1
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
viivakoodi==0.8.0
